requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
packaging==26.0
pydantic==2.12.5
pydantic_core==2.41.5
pyisckea==0.9.2
pykeadhcp==0.6.0
pynetbox==7.6.1
python-dotenv==1.2.1
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
